Cache textures by content hash instead of CPU pointer.
Pointer-keyed caching caused false evictions when the game reused the same
memory address for different texture data. Replaced the uintptr_t key with
a fast FNV-1a sample hash (start + middle + end of the 16-bit pixel buffer
plus dimensions, ~100 cycles constant cost) so identical content always hits
regardless of pointer identity, and changed content always misses.

Also removed a leftover debug fopen() in UploadTexture16 that wrote to a
hardcoded absolute path, and early-exit the transparency scan once any
transparent pixel is found.

+++ renderer/RendererGL.h
@@ -141,12 +141,12 @@ private:
     GLuint m_vao = 0;
     GLuint m_vbo = 0;
 
-    // Texture cache: maps (cpuAddr, w, h) → GL texture ID
+    // Texture cache: maps content hash → GL texture ID
     struct TexCacheEntry {
         GLuint texId;
         int lastUsed;
     };
-    std::unordered_map<uintptr_t, TexCacheEntry> m_texCache;
+    std::unordered_map<uint64_t, TexCacheEntry> m_texCache;
     GLuint m_currentTexture = 0;
     int    m_frameCounter = 0;
